requirements.txt
asgiref==3.2.10
Django==3.0.8
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.6.0
django-crispy-forms==1.9.2
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
Pillow==7.2.0
progressbar2==3.51.4
PyJWT==1.7.1
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
